rgkk-1.labos/rgkk-1.labos/barbara-pavlović-0036526060/requirements.txt
python-bitcoinlib
requests~=2.31.0

pip~=23.3.1
Pillow~=9.4.0
certifi~=2023.7.22
idna~=3.4
urllib3~=2.0.7
setuptools~=65.5.0
packaging~=23.0
pyparsing~=3.0.9